more DeleteInstance (#810)
* more DeleteInstance

* Update Wrapper.cs

+++ src/LocalDb/Wrapper.cs
@@ -231,7 +231,7 @@ class Wrapper : IDisposable
     public void DeleteInstance(ShutdownMode mode = ShutdownMode.KillProcess)
     {
         LocalDbApi.StopAndDelete(instance, mode);
-        System.IO.Directory.Delete(Directory, true);
+        DeleteDirectory();
         Dispose();
     }
 
@@ -239,8 +239,16 @@ class Wrapper : IDisposable
     public void DeleteInstance(ShutdownMode mode, TimeSpan timeout)
     {
         LocalDbApi.StopAndDelete(instance, mode, timeout);
-        System.IO.Directory.Delete(Directory, true);
+        DeleteDirectory();
         Dispose();
+    }
+
+    void DeleteDirectory()
+    {
+        if (System.IO.Directory.Exists(Directory))
+        {
+            System.IO.Directory.Delete(Directory, true);
+        }
     }
 
     void DeleteTemplateFiles()
